Refactor autoProcessTV -> tv
